--- .hintrc
@@ -1,8 +1,0 @@
-{
-  "extends": [
-    "development"
-  ],
-  "hints": {
-    "disown-opener": "off"
-  }
-}
