+++ api/package.json
@@ -5,8 +5,7 @@
   "dependencies": {
     "@redwoodjs/api": "2.2.0",
     "@redwoodjs/api-server": "^2.2.0",
-    "@redwoodjs/graphql-server": "2.2.0",
-    "date-fns": "^2.29.3"
+    "@redwoodjs/graphql-server": "2.2.0"
   },
   "devDependencies": {
     "@types/chance": "^1.1.3",
