+++ web/package.json
@@ -19,7 +19,8 @@
     "@redwoodjs/web": "2.2.0",
     "prop-types": "15.8.1",
     "react": "17.0.2",
-    "react-dom": "17.0.2"
+    "react-dom": "17.0.2",
+    "date-fns": "^2.29.3"
   },
   "devDependencies": {
     "@playwright/test": "^1.24.2",
