+++ web/src/components/Footer/Footer.stories.tsx
@@ -1,0 +1,7 @@
+import { Footer } from './Footer'
+
+export const generated = (args) => {
+  return <Footer {...args} />
+}
+
+export default { title: 'Components/Footer' }
